Extract nerve config parsing to its own file
