+++ src/fullerite/collector/nerve_uwsgi.go
@@ -1,13 +1,12 @@
 package collector
 
 import (
-	"fullerite/config"
 	"fullerite/metric"
+	"fullerite/util"
 
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
-	"net"
 	"net/http"
 	"strconv"
 	"strings"
@@ -28,25 +27,6 @@ const (
 // all the local endpoints on the given path. The metrics are assumed
 // to be in a UWSGI format.
 //
-// example configuration::
-//
-// {
-//     	"heartbeat_path": "/var/run/nerve/heartbeat",
-//		"instance_id": "srv1-devc",
-//		"services": {
-//	 		"<SERVICE_NAME>.<otherstuff>": {
-//				"host": "<IPADDR>",
-//      	    "port": ###,
-//      	}
-// 		}
-//     "services": {
-//
-// Most imporantly is the port, host and service name. The service name is assumed to be formatted like this::
-//
-type releventNerveConfig struct {
-	Services map[string]map[string]interface{}
-}
-
 // the assumed format is:
 // {
 // 	"gauges": {},
@@ -126,19 +106,13 @@ type nerveUWSGICollector struct {
 }
 
 func (n *nerveUWSGICollector) Collect() {
-	ips, err := getIps()
-	if err != nil {
-		n.log.Warn("Failed to get IPs: ", err)
-		return
-	}
-
 	rawFileContents, err := ioutil.ReadFile(n.configFilePath)
 	if err != nil {
 		n.log.Warn("Failed to read the contents of file ", n.configFilePath, " because ", err)
 		return
 	}
 
-	servicePortMap, err := n.parseNerveConfig(&rawFileContents, ips)
+	servicePortMap, err := util.ParseNerveConfig(&rawFileContents)
 	if err != nil {
 		n.log.Warn("Failed to parse the nerve config at ", n.configFilePath, ": ", err)
 		return
@@ -187,41 +161,6 @@ func (n *nerveUWSGICollector) Configure(configMap map[string]interface{}) {
 	}
 
 	n.configureCommonParams(configMap)
-}
-
-// parseNerveConfig is responsible for taking the JSON string coming in into a map of service:port
-// it will also filter based on only services runnign on this host.
-// To deal with multi-tenancy we actually will return port:service
-func (n *nerveUWSGICollector) parseNerveConfig(raw *[]byte, ips []string) (map[int]string, error) {
-	parsed := new(releventNerveConfig)
-
-	// convert the ips into a map for membership tests
-	ipMap := make(map[string]bool)
-	for _, val := range ips {
-		ipMap[val] = true
-	}
-
-	err := json.Unmarshal(*raw, parsed)
-	if err != nil {
-		return make(map[int]string), err
-	}
-	results := make(map[int]string)
-	for rawServiceName, serviceConfig := range parsed.Services {
-		host := strings.TrimSpace(serviceConfig["host"].(string))
-
-		_, exists := ipMap[host]
-		if exists {
-			name := strings.Split(rawServiceName, ".")[0]
-			port := config.GetAsInt(serviceConfig["port"], -1)
-			if port != -1 {
-				results[port] = name
-			} else {
-				n.log.Warn("Failed to get port from value ", serviceConfig["port"])
-			}
-		}
-	}
-
-	return results, nil
 }
 
 // parseDefault is the fallback parser if no 'Metrics-Schema' is provided in the
@@ -357,35 +296,6 @@ func createMetricFromDatam(rollup string, value interface{}, metricName string, 
 		return m, false
 	}
 	return m, true
-}
-
-// getIps is responsible for getting all the ips that are associated with this NIC
-func getIps() ([]string, error) {
-	ifaces, err := net.Interfaces()
-	if err != nil {
-		return []string{}, err
-	}
-
-	results := []string{}
-	for _, i := range ifaces {
-		addrs, err := i.Addrs()
-		if err != nil {
-			return []string{}, err
-		}
-		for _, addr := range addrs {
-			var ip net.IP
-			switch v := addr.(type) {
-			case *net.IPNet:
-				ip = v.IP
-			case *net.IPAddr:
-				ip = v.IP
-			}
-
-			results = append(results, ip.String())
-		}
-	}
-
-	return results, nil
 }
 
 func queryEndpoint(endpoint string, timeout int) ([]byte, string, error) {
